- added "Host group" element to maps (Eugene)
git-svn-id: svn://svn.zabbix.com/trunk@3499 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ frontends/php/include/hosts.inc.php
@@ -328,6 +328,9 @@ require_once "include/items.inc.php";
 
 	function	delete_host_group($groupid)
 	{
+		if(!delete_sysmaps_elements_with_groupid($groupid))
+			return false;
+		
 		if(!DBexecute("delete from hosts_groups where groupid=$groupid"))
 			return false;
